pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "squaremap_combine"
version = "0.21.0"
authors = [
  { name="Seth 'Violet' Gibbs" },
]
description = "Provides tools to combine tile images created by the 'squaremap' Minecraft plugin into one single world image."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies=[
  "loguru==0.7.2",
  "pillow==10.4.0",
  "tqdm==4.66.5"
]

[project.optional-dependencies]
gui = [
  "dearpygui==1.11.1",
  "pydantic==2.9.2"
]

[project.scripts]
squaremap-combine = "squaremap_combine.combine_cli:main"

[project.gui-scripts]
squaremap-combine-gui = "squaremap_combine.gui.combine_gui:main"

[project.urls]
Homepage = "https://github.com/svioletg/py-squaremap-combiner"
Issues = "https://github.com/svioletg/py-squaremap-combiner/issues"